Rename operations to arithmetic
