+++ benchmarks/CMakeLists.txt
@@ -37,6 +37,6 @@ function(_create_benchmark TEST_FILE)
 endfunction()
 
 _create_benchmark("algorithms")
-_create_benchmark("operations")
+_create_benchmark("arithmetic")
 _create_benchmark("initialization")
 _create_benchmark("utilities")
